Use Angular style-guide
