+++ package.json
@@ -44,7 +44,6 @@
     "style-loader": "^0.13.1",
     "typescript": "^1.8.10",
     "webpack": "^1.13.0",
-    "webpack-dashboard": "0.0.1",
     "webpack-dev-server": "^1.14.1"
   },
   "dependencies": {
@@ -55,9 +54,9 @@
     "@angular/platform-browser": "2.0.0-rc.5",
     "@angular/platform-browser-dynamic": "2.0.0-rc.5",
     "@angular/router": "3.0.0-rc.1",
-    "angular2-apollo": "^0.4.2",
+    "angular2-apollo": "^0.4.4",
     "angular2-moment": "^0.8.0",
-    "apollo-client": "^0.4.11",
+    "apollo-client": "^0.4.12",
     "apollo-server": "^0.2.5",
     "body-parser": "^1.15.1",
     "classnames": "^2.2.5",
@@ -81,7 +80,7 @@
     "redux": "^3.5.2",
     "reflect-metadata": "^0.1.3",
     "request-promise": "^4.1.1",
-    "rxjs": "5.0.0-beta.11",
+    "rxjs": "5.0.0-beta.6",
     "sqlite3": "^3.1.4",
     "ts-loader": "^0.8.2",
     "whatwg-fetch": "^1.0.0",
