+++ ui/app/app.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app',
+  template: `
+    <navigation></navigation>
+    <div class="container">
+      <router-outlet></router-outlet>
+    </div>
+  `
+})
+export class AppComponent {}
