+++ ui/app/comments/index.ts
@@ -1,0 +1,13 @@
+import { RouterConfig } from '@angular/router';
+
+import { CommentComponent } from './comment.component';
+import { CommentsPageComponent } from './comments-page.component';
+
+export const COMMENTS_DECLARATIONS = [
+  CommentComponent,
+  CommentsPageComponent
+];
+
+export const COMMENTS_ROUTES: RouterConfig = [
+  { path: ':org/:repoName', component: CommentsPageComponent }
+];
