+++ ui/app/feed/index.ts
@@ -1,0 +1,16 @@
+import { RouterConfig } from '@angular/router';
+
+import { FeedComponent } from './feed.component';
+import { FeedEntryComponent } from './feed-entry.component';
+import { VoteButtonsComponent } from './vote-buttons.component';
+
+export const FEED_DECLARATIONS = [
+  FeedComponent,
+  FeedEntryComponent,
+  VoteButtonsComponent
+];
+
+export const FEED_ROUTES: RouterConfig = [
+  { path: '', redirectTo: '/feed/top', pathMatch: 'full' },
+  { path: 'feed/:type', component: FeedComponent }
+];
