+++ ui/app/shared/index.ts
@@ -1,0 +1,11 @@
+import { EmojifyPipe } from './emojify.pipe';
+import { InfoLabelComponent } from './info-label.component';
+import { LoadingComponent } from './loading.component';
+import { RepoInfoComponent } from './repo-info.component';
+
+export const SHARED_DECLARATIONS = [
+  EmojifyPipe,
+  InfoLabelComponent,
+  LoadingComponent,
+  RepoInfoComponent
+];
