+++ ui/app/shared/info-label.component.ts
@@ -1,0 +1,12 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'info-label',
+  template: `
+    <span class="label label-info">{{label}}: {{value}}</span>
+  `
+})
+export class InfoLabelComponent {
+  @Input() label;
+  @Input() value;
+}
