+++ ui/index.html
@@ -18,7 +18,7 @@
     </script>
   </head>
   <body>
-    <git-hunt></git-hunt>
+    <app></app>
     <div id="footer">
       <ul>
         <li>Fork on <a href="https://github.com/kamilkisiela/GitHunt-angular2">Github</a></li>
