+++ ui/index.ts
@@ -1,13 +1,11 @@
-import 'es6-shim';
 import 'reflect-metadata';
+import 'rxjs';
 import 'zone.js/dist/zone';
-
+import 'zone.js/dist/long-stack-trace-zone';
 import './style.css';
 
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import { AppModule }              from './app.module';
 
-platformBrowserDynamic().bootstrapModule(AppModule).catch((error) => {
-    console.log('error', error);
-});
+import { AppModule } from './app/app.module';
 
+platformBrowserDynamic().bootstrapModule(AppModule);
