+++ webpack.config.js
@@ -1,7 +1,4 @@
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-var Dashboard = require('webpack-dashboard');
-var DashboardPlugin = require('webpack-dashboard/plugin');
-var dashboard = new Dashboard();
 
 module.exports = {
   entry: './ui/index.ts',
@@ -28,7 +25,6 @@ module.exports = {
     new HtmlWebpackPlugin({
       template: 'ui/index.html',
     }),
-    new DashboardPlugin(dashboard.setData),
   ],
   devServer: {
     proxy: {
